requirements.txt
streamlit
pandas
plotly
folium
streamlit-folium
Pillow
requests
koreanize-matplotlib
